# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.rachitpatel.com/v1",
            "description": "Live Base URL"
        },
        {
            "url": "https://api-laravel.test/v1",
            "description": "Base URL"
        },
        {
            "url": "https://dp-api.local/v1",
            "description": "Base URL 1"
        },
        {
            "url": "http://127.0.0.1:8000/v1",
            "description": "Local Base URL"
        }
    ],
    "paths": {
...
        "/role/": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get All roles",
                "description": "Display all the roles",
                "operationId": "showRoles",
                "requestBody": {},
                "responses": {
                    "204": {
                        "description": "Roles data show successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Roles data get successfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 204
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Create a new role",
                "description": "Create a new role with the provided information.",
                "operationId": "2d421f2bcc698508500668000572ad3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Role's name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "string",
                                            "example": "201"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/role/{id}": {
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Update role",
                "description": "Updates the role's name",
                "operationId": "updateRoleProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role updated successfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 204
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleUpdateRequest": {
                "title": "Role Update Request",
                "description": "Request data for updating role.",
                "required": [
                    "name",
                    "admin"
                ],
                "properties": {
                    "name": {
                        "description": "Role's name",
                        "type": "string",
                        "example": "Admin"
                    }
                },
                "type": "object"
            },
...
            "RoleResource": {
                "title": "Role Resource",
                "description": "Role resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "500": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Oops something went wrong"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 500
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "401": {
                "description": "Forbidden: Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 401
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Validation Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Validation error"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "field_name": [
                                            "The field_name field is required."
                                        ]
                                    }
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 422
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}